Feat: Car 객체 구현
